Extracted SelectPaymentMethod screen to a separate package
