+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/SelectPaymentMethodViewModelTest.kt
@@ -11,8 +11,6 @@ import com.woocommerce.android.model.Order
 import com.woocommerce.android.model.OrderMapper
 import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
-import com.woocommerce.android.ui.payments.SelectPaymentMethodViewState.Loading
-import com.woocommerce.android.ui.payments.SelectPaymentMethodViewState.Success
 import com.woocommerce.android.ui.payments.cardreader.CardReaderTracker
 import com.woocommerce.android.ui.payments.cardreader.LearnMoreUrlProvider
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderFlowParam
@@ -24,12 +22,14 @@ import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderFlowP
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderFlowParam.PaymentOrRefund.Refund
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderType
 import com.woocommerce.android.ui.payments.cardreader.payment.CardReaderPaymentCollectibilityChecker
-import com.woocommerce.android.ui.payments.NavigateBackToHub
-import com.woocommerce.android.ui.payments.NavigateBackToOrderList
-import com.woocommerce.android.ui.payments.NavigateToCardReaderHubFlow
-import com.woocommerce.android.ui.payments.NavigateToCardReaderRefundFlow
-import com.woocommerce.android.ui.payments.NavigateToOrderDetails
-import com.woocommerce.android.ui.payments.SelectPaymentMethodViewModel
+import com.woocommerce.android.ui.payments.methodselection.NavigateBackToHub
+import com.woocommerce.android.ui.payments.methodselection.NavigateBackToOrderList
+import com.woocommerce.android.ui.payments.methodselection.NavigateToCardReaderHubFlow
+import com.woocommerce.android.ui.payments.methodselection.NavigateToCardReaderRefundFlow
+import com.woocommerce.android.ui.payments.methodselection.NavigateToOrderDetails
+import com.woocommerce.android.ui.payments.methodselection.SelectPaymentMethodViewModel
+import com.woocommerce.android.ui.payments.methodselection.SelectPaymentMethodViewState.Loading
+import com.woocommerce.android.ui.payments.methodselection.SelectPaymentMethodViewState.Success
 import com.woocommerce.android.ui.payments.taptopay.IsTapToPayAvailable
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.util.captureValues
